refactor: move npm.ts to subdirectory
